authorization code grant type : Unit Test to write

+++ DaOAuthV2.Service.DTO/OAuth/TokenInfoDto.cs
@@ -4,7 +4,7 @@
     {
         public string AccessToken { get; set; }
         public string TokenType { get; set; }
-        public int ExpireIn { get; set; }
+        public long ExpireIn { get; set; }
         public string RefreshToken { get; set; }
         public string Scope { get; set; }
     }
